Clean up model builder API, add a cube helper, update examples
