+++ src/main/resources/META-INF/accesstransformer.cfg
@@ -37,6 +37,7 @@ public net.minecraft.client.renderer.model.BlockModel field_178315_d # parent
 public net.minecraft.client.renderer.model.BlockModel field_178318_c # textures
 public net.minecraft.client.renderer.model.BlockModel field_178322_i # ambientOcclusion
 public net.minecraft.client.renderer.model.BlockModel func_187966_f()Ljava/util/List; # getOverrides
+public net.minecraft.client.renderer.model.BlockPart func_178236_a(Lnet/minecraft/util/Direction;)[F # getFaceUvs
 protected net.minecraft.client.renderer.model.ItemOverrideList <init>()V
 protected net.minecraft.client.renderer.model.ModelBakery field_177598_f # resourceManager
 protected net.minecraft.client.renderer.model.ModelBakery field_177602_b # LOCATIONS_BUILTIN_TEXTURES
